app/requirements.txt
dj-database-url==0.4.1
Django==1.11.1
django-forms-bootstrap==3.1.0
django-watson==1.4.0
gunicorn==19.6.0
psycopg2==2.6.2
pytz==2017.2
whitenoise==3.2
